Make the kubeseal CLI have a skinny main
